remove file from server

+++ controllers/file/deleteFile.js
@@ -1,19 +1,49 @@
 import File from '../../models/file.js';
+import fs from 'fs/promises';
+
+const removeFileFromPublicFolder = async (path, res) => {
+    try {
+        await fs.unlink(path);
+        return true;
+    } catch (err) {
+        res.status(500).send({
+            success: false,
+            status: res.status,
+            message: err.message
+        });
+        return false;
+    }
+}
+
 const deleteFile = async (req, res) => {
     try {
-        const file = await File.destroy({
+        const file = await File.findOne({
             where: {
                 user_id: req.userData.userId,
                 file_id: req.params.id
             }
         })
-        if(file) {
-            res.status(200).send({
-                success: true,
-                status: res.status,
-                message: `Delete file with id ${req.params.id}`
-            });
-        } else {
+
+        if (file) {
+            const removed = await removeFileFromPublicFolder(file.path, res);
+
+            if (removed) {
+
+                await File.destroy({
+                    where: {
+                        file_id: file.file_id
+                    }
+                });
+
+                res.status(200).send({
+                    success: true,
+                    status: res.status,
+                    message: `Delete file with id ${req.params.id}`
+                });
+            }
+
+        }
+        else {
             res.status(400).send({
                 success: false,
                 status: res.status,
@@ -21,7 +51,7 @@ const deleteFile = async (req, res) => {
             });
         }
     }
-    catch(err) {
+    catch (err) {
         res.status(500).send({
             success: false,
             status: res.status,
